% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{create_subtree}
\alias{create_subtree}
\title{Create a subtree dataframe}
\usage{
create_subtree(parents, specs, level)
}
\arguments{
\item{parents}{Integer() with parent's ids.}

\item{specs}{List of tree specs.}

\item{level}{The tree level.}
}
\value{
Dataframe with subtree.
}
\description{
Create a subtree dataframe.
}
\details{
Create a subtree dataframe used to build a tree with integer as id.
}
\examples{
root <- 1L
specs <- list(
 l1 = list(digits = 100, nnodes = 9),
 l2 = list(digits = 100, nnodes = 19),
 l3 = list(digits = 100, nnodes = 29))
df <- data.frame(
 parent = root,
 child = root,
 level = 0L)
 out <- create_subtree(parents = df$child, specs[[1]], level = 1)
stopifnot((nrow(out) == specs[[1]]$nnodes))
}
