% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{create_tree}
\alias{create_tree}
\title{Create a tree with integer id}
\usage{
create_tree(specs, root = 1L)
}
\arguments{
\item{specs}{List of tree specs.}

\item{root}{Integer identifying the integer, default is 1.}
}
\value{
Dataframe with tree.
}
\description{
Create a tree with integer id.
}
\details{
Create a tree with integer id and level in a third column.
}
\examples{
specs <- list(
 l1 = list(digits = 100, nnodes = 9),
 l2 = list(digits = 100, nnodes = 19),
 l3 = list(digits = 100, nnodes = 29))
create_tree(specs)
}
