% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflTools-package.R
\docType{package}
\name{eflTools}
\alias{eflTools}
\alias{eflTools-package}
\title{A package of tools used by FL}
\description{
Funcitons to facilitate the analysis performed by FL
}
\section{eflTools functions}{

The functions ...
}

\author{
\strong{Maintainer}: François Lefebvre \email{flefebvre01@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
