% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_acctg.R
\name{exp_acctg}
\alias{exp_acctg}
\title{Signed Exponentiation With a Base}
\usage{
exp_acctg(x, base = 10L)
}
\arguments{
\item{x}{Numerical vector.}

\item{base}{Positive number >= 2. Default value is 10.}
}
\value{
Numeric vector.
}
\description{
Signed exponentiation With a base.
}
\details{
Compute \code{exp_acctg(x, base) = sign(x) * (base ^ abs(x) - 1)}.
This function is the inverse of \code{log_acctg}.
}
\examples{
x <- c(-2, -1, -1/10, -0.1/10, 0, 0.1/10, 1/10, 1, 2)
y <- exp_acctg(x)
# this test if the inverse is ok
stopifnot(isTRUE(all.equal(x, eflTools::log_acctg(y))))
}
\seealso{
log_acctg
}
