% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1ps.R
\name{expm1s}
\alias{expm1s}
\title{Signed \code{expm1} With a Base}
\usage{
expm1s(x, base = exp(1))
}
\arguments{
\item{x}{Numerical vector.}

\item{base}{Positive number >= 2. Default value is exp(1).}
}
\value{
Numeric vector.
}
\description{
Signed \code{expm1} With a base.
}
\details{
Compute \code{expm1s(x, base) = sign(x) * (base ^ abs(x) - 1)}.
This function is the inverse of \code{log1ps}.
}
\examples{
x <- c(-2, -1, -1/10, -0.1/10, 0, 0.1/10, 1/10, 1, 2)
y <- expm1s(x, base = 10)
# this test if the inverse is ok
stopifnot(isTRUE(all.equal(x, eflTools::log1ps(y, base = 10))))
}
\seealso{
log1ps
}
