% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1ps.R
\name{expm1s10}
\alias{expm1s10}
\title{Signed \code{expm1} With Base 10}
\usage{
expm1s10(x)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
Numeric vector.
}
\description{
Signed \code{expm1} With base 10.
}
\details{
Compute \code{expm1s10(x) = sign(x) * (10 ^ abs(x) - 1)}.
This function is the inverse of \code{log1ps10}.
}
\examples{
x <- c(-2, -1, -1/10, -0.1/10, 0, 0.1/10, 1/10, 1, 2)
y <- expm1s10(x)
# this test if the inverse is ok
stopifnot(isTRUE(all.equal(x, eflTools::log1ps10(y))))
}
\seealso{
expm1s log1ps10
}
