% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_acctg_tags.R
\name{gather_acctg_tags}
\alias{gather_acctg_tags}
\title{Import acctg tags from an excel file}
\usage{
gather_acctg_tags(
  file,
  sheet = "tbl_imp_tags",
  cols = c("concept", "entity", "fsstatus", "curr_no")
)
}
\arguments{
\item{file}{Path and file name of the excel file.}

\item{sheet}{Name of the excel sheet.}

\item{cols}{Character vector with the names of the columns not pivoted.}
}
\value{
Dataframe.
}
\description{
Import acctg tags from an excel file.
}
\details{
Import acctg tags from an excel file in a long format. Usually to be exported
to MS Access.
}
