% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{get_bag.bobj}
\alias{get_bag.bobj}
\title{Get Bag Item from a Business Object}
\usage{
\method{get_bag}{bobj}(x, name)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{name}{Name of the bag item.}
}
\value{
Bag from an object of class \emph{bobj}.
}
\description{
Get Bag Item from a Business Object
}
