% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_bag.default}
\alias{get_bag.default}
\title{Get Bag Default Method}
\usage{
\method{get_bag}{default}(x, name)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of bag item.}
}
\value{
Data from an object.
}
\description{
Get Bag Default Method
}
