% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{get_bag.xprts}
\alias{get_bag.xprts}
\title{Get Bag Item from an Export Object}
\usage{
\method{get_bag}{xprts}(x, name)
}
\arguments{
\item{x}{Object of class \emph{xprts}.}

\item{name}{Name of the bag item.}
}
\value{
Bag from an object of class \emph{xprts}.
}
\description{
Get Bag Item from an Export Object
}
