% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{get_data.bobj}
\alias{get_data.bobj}
\title{Get Data from a Business Object}
\usage{
\method{get_data}{bobj}(x)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}
}
\value{
Data from an object of class \emph{bobj}.
}
\description{
Get Data from a Business Object
}
