% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_data.default}
\alias{get_data.default}
\title{Get Data Default Method}
\usage{
\method{get_data}{default}(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Data from an object.
}
\description{
Get Data Default Method
}
