% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{get_formula.bobj}
\alias{get_formula.bobj}
\title{Get Formula from a Business Object}
\usage{
\method{get_formula}{bobj}(x)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}
}
\value{
Formula from an object of class \emph{bobj}.
}
\description{
Get Formula from a Business Object
}
