% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_formula.default}
\alias{get_formula.default}
\title{Get Formula Default Method}
\usage{
\method{get_formula}{default}(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Data from an object.
}
\description{
Get Formula Default Method
}
