% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{get_info.bobj}
\alias{get_info.bobj}
\title{Get Info Item from a Business Object}
\usage{
\method{get_info}{bobj}(x, name)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{name}{Name of the info item.}
}
\value{
Info from an object of class \emph{bobj}.
}
\description{
Get Info Item from a Business Object
}
