% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_info.default}
\alias{get_info.default}
\title{Get Info Default Method}
\usage{
\method{get_info}{default}(x, name)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of info item.}
}
\value{
Data from an object.
}
\description{
Get Info Default Method
}
