% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{get_roles.bobj}
\alias{get_roles.bobj}
\title{Get Roles from a Business Object}
\usage{
\method{get_roles}{bobj}(x)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}
}
\value{
Roles from an object of class \emph{bobj}.
}
\description{
Get Roles from a Business Object
}
