% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_roles.default}
\alias{get_roles.default}
\title{Get Roles Default Method}
\usage{
\method{get_roles}{default}(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Data from an object.
}
\description{
Get Roles Default Method
}
