% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{init_roles}
\alias{init_roles}
\title{Initialize Roles of a Business Object}
\usage{
init_roles(data, formula, id)
}
\arguments{
\item{data}{Dataframe.}

\item{formula}{Formula of business model.}

\item{id}{Character vector of key variables.}
}
\value{
Dataframe of roles.
}
\description{
Initialize Roles of a Business Object
}
