% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_access.R
\name{load_access}
\alias{load_access}
\title{Load data from an MS Access db}
\usage{
load_access(dir = getwd(), db = NULL, qries = NULL, env = .GlobalEnv)
}
\arguments{
\item{dir}{directory name of the MS Access db}

\item{db}{file name of the MS Access db}

\item{qries}{query names in the form "df_name" = "qry_name"
where df_name is the name that will be assigned to the dataframe
and qry_name is the query name in the db}

\item{env}{the environment where the data will be located}
}
\value{
nb of queries processed
}
\description{
Load data from an MS Access db
}
\details{
Extract the data from a local MS Access database using a list of queries.
}
\examples{
load_access(
  dir = system.file("extdata", package = "eflTools"),
  db = "db_MsAccess.accdb",
  qries = c("df_clients" = "tbl_clients", "df_vendors" = "tbl_vendors"))
}
