% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1ps.R
\name{log1ps}
\alias{log1ps}
\title{Signed \code{log1p} With a Base}
\usage{
log1ps(x, base = exp(1))
}
\arguments{
\item{x}{Numerical vector.}

\item{base}{Positive number >= 2. Default value is exp(1).}
}
\value{
Numeric vector.
}
\description{
Signed \code{log1p} with a base.
}
\details{
Compute \code{log1ps(x, base) = sign(x) * log1p(abs(x)) / log(base)}.
This function is the inverse of \code{expm1s}.
}
\examples{
x <- c(-100, -10, -1, -0.1, 0, 0.1, 1, 10, 100)
y <- log1ps(x, base = 10)
# this test if the inverse is ok
stopifnot(isTRUE(all.equal(x, eflTools::expm1s(y, base = 10))))
}
\seealso{
expm1s
}
