% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1ps.R
\name{log1ps10}
\alias{log1ps10}
\title{Signed \code{log1p} With Base 10}
\usage{
log1ps10(x)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
Numeric vector.
}
\description{
Signed \code{log1p} with base 10.
}
\details{
Compute \code{log1ps10(x) = sign(x) * log1p(abs(x)) / log(10)}.
This function is the inverse of \code{expm1s10}.
}
\examples{
x <- c(-100, -10, -1, -0.1, 0, 0.1, 1, 10, 100)
y <- log1ps10(x)
# this test if the inverse is ok
stopifnot(isTRUE(all.equal(x, eflTools::expm1s10(y))))
}
\seealso{
log1ps expm1s10
}
