% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_acctg.R
\name{log_acctg}
\alias{log_acctg}
\title{Log Using Accounting Rule}
\usage{
log_acctg(x, base = 10L)
}
\arguments{
\item{x}{Numerical vector.}

\item{base}{Positive number >= 2.}
}
\value{
Numeric vector.
}
\description{
Log using accounting rule.
}
\details{
Compute the log for number whose absolute value is greater than of equal to
the base and change the sign of the result to be the same as the input. If
the absolute value of the number is less than the base, then divide it
by the base.
}
\examples{
x <- c(-100, -10, -1, -0.1, 0, 0.1, 1, 10, 100)
y <- log_acctg(x)
stopifnot(identical(y, c(-2, -1, -1/10, -0.1/10, 0, 0.1/10, 1/10, 1, 2)))
}
\seealso{
exp_acctg
}
