% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{new_bobj}
\alias{new_bobj}
\title{Create New Business Object}
\usage{
new_bobj(data, formula, id)
}
\arguments{
\item{data}{Dataframe.}

\item{formula}{Formula of business model.}

\item{id}{Character vector of key variables.}
}
\value{
Object of class \emph{bobj}.
}
\description{
Create New Business Object
}
