% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{new_xprts}
\alias{new_xprts}
\title{Create New Export Object of class \code{xprts}}
\usage{
new_xprts()
}
\value{
Object of class \emph{xprts}.
}
\description{
Create New Export Object of class \code{xprts}
}
