% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{print.bobj}
\alias{print.bobj}
\title{Print a Business Object of Class \code{bobj}}
\usage{
\method{print}{bobj}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Formatted message to console.
}
\description{
Print a Business Object of Class \code{bobj}
}
