% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{print.xprts}
\alias{print.xprts}
\title{Print an Export Object of class \code{xprts}}
\usage{
\method{print}{xprts}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{xprts}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Formatted message to console.
}
\description{
Print an Export Object of class \code{xprts}
}
