% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eda.R
\name{run_eda}
\alias{run_eda}
\title{Run R scripts in sequence}
\usage{
run_eda(dir = file.path(getwd(), "src"), pattern = "^eda.*[.]R$")
}
\arguments{
\item{dir}{String. Name of the directory where the R scripts are located.}

\item{pattern}{Regex pattern. Pattern to filter the source files to run.}
}
\value{
Number of source files run.
}
\description{
Run R scripts in sequence
}
\details{
Run R scripts named after a pattern in a specified directory.
}
\examples{
\dontrun{
run_eda(dir = file.path(getwd(), "src"), pattern = "^eda.*[.]R$")
}
}
