% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eda.R
\name{run_eda_msg}
\alias{run_eda_msg}
\title{Console message when run_eda() is completed successfully}
\usage{
run_eda_msg(nb)
}
\arguments{
\item{nb}{number of scripts successfully completed}
}
\value{
output to console
}
\description{
Console message when run_eda() is completed successfully
}
\details{
This is used by run_eda() and not exported.
}
\examples{
run_eda_msg(1)
}
