% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{set_bag.bobj}
\alias{set_bag.bobj}
\title{Set Bag Item from a Business Object}
\usage{
\method{set_bag}{bobj}(x, name, value)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{name}{Name of the bag item.}

\item{value}{New value to assign to the bag.}
}
\value{
Object of class \emph{bobj}.
}
\description{
Set Bag Item from a Business Object
}
