% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_bag.default}
\alias{set_bag.default}
\title{Set Bag Default Method}
\usage{
\method{set_bag}{default}(x, name, value)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of bag item.}

\item{value}{Value of bag item.}
}
\value{
Data from an object.
}
\description{
Set Bag Default Method
}
