% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_bag}
\alias{set_bag}
\title{Set Bag of a Object}
\usage{
set_bag(x, name, value)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of bag item.}

\item{value}{Value of bag item.}
}
\description{
Set Bag of a Object
}
