% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{set_bag.xprts}
\alias{set_bag.xprts}
\title{Set Bag Item from an Export Object}
\usage{
\method{set_bag}{xprts}(x, name, value)
}
\arguments{
\item{x}{Object of class \emph{xprts}.}

\item{name}{Name of the bag item.}

\item{value}{New value to assign to the bag.}
}
\value{
Object of class \emph{xprts}.
}
\description{
Set Bag Item from an Export Object
}
