% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{set_data.bobj}
\alias{set_data.bobj}
\title{Set Data from a Business Object}
\usage{
\method{set_data}{bobj}(x, data)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{data}{Dataframe.}
}
\value{
Data from an object of class \emph{bobj}.
}
\description{
Set Data from a Business Object
}
