% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_data.default}
\alias{set_data.default}
\title{Set Data Default Method}
\usage{
\method{set_data}{default}(x, data)
}
\arguments{
\item{x}{Object.}

\item{data}{Dataframe.}
}
\value{
Data from an object.
}
\description{
Set Data Default Method
}
