% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{set_formula.bobj}
\alias{set_formula.bobj}
\title{Set Formula from a Business Object}
\usage{
\method{set_formula}{bobj}(x, formula)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{formula}{Formula.}
}
\value{
Data from an object of class \emph{bobj}.
}
\description{
Set Formula from a Business Object
}
