% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_formula.default}
\alias{set_formula.default}
\title{Set Formula Default Method}
\usage{
\method{set_formula}{default}(x, formula)
}
\arguments{
\item{x}{Object.}

\item{formula}{Formula.}
}
\value{
Data from an object.
}
\description{
Set Formula Default Method
}
