% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{set_info.bobj}
\alias{set_info.bobj}
\title{Set Info Item from a Business Object}
\usage{
\method{set_info}{bobj}(x, name, value)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{name}{Name of the info item.}

\item{value}{New value to assign to the info.}
}
\value{
Object of class \emph{bobj}.
}
\description{
Set Info Item from a Business Object
}
