% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_info.default}
\alias{set_info.default}
\title{Set Info Default Method}
\usage{
\method{set_info}{default}(x, name, value)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of the info item.}

\item{value}{Value of the info item.}
}
\value{
Data from an object.
}
\description{
Set Info Default Method
}
