% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_info}
\alias{set_info}
\title{Set Info of a Object}
\usage{
set_info(x, name, value)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of the info item.}

\item{value}{Value of the info item.}
}
\description{
Set Info of a Object
}
