% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{set_roles.bobj}
\alias{set_roles.bobj}
\title{Set Roles from a Business Object}
\usage{
\method{set_roles}{bobj}(x, roles)
}
\arguments{
\item{x}{Object of class \emph{bobj}.}

\item{roles}{New values to assign to the roles.}
}
\value{
Object of class \emph{bobj}.
}
\description{
Set Roles from a Business Object
}
