% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_roles.default}
\alias{set_roles.default}
\title{Set Roles Default Method}
\usage{
\method{set_roles}{default}(x, roles)
}
\arguments{
\item{x}{Object.}

\item{roles}{Dataframe of roles.}
}
\value{
Data from an object.
}
\description{
Set Roles Default Method
}
