% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobj.R
\name{validate_bobj}
\alias{validate_bobj}
\title{Validate Business Object of class \code{bobj}}
\usage{
validate_bobj(obj)
}
\arguments{
\item{obj}{Object of class \emph{bobj}.}
}
\value{
Object of class \emph{bobj}.
}
\description{
Validate Business Object of class \code{bobj}
}
