% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{validate_xprts}
\alias{validate_xprts}
\title{Validate Export Object of class \code{xprts}}
\usage{
validate_xprts(obj)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}
}
\value{
Object of class \emph{xprts}.
}
\description{
Validate Export Object of class \code{xprts}
}
