% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{xprt_all}
\alias{xprt_all}
\title{Export all objects from Export Object of class \code{xprts}}
\usage{
xprt_all(obj, path, is_xprt)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path for exported file.}

\item{is_xprt}{TRUE = Export objects, FALSE = Don't export.}
}
\value{
\code{obj} invisibly.
}
\description{
Export all objects from Export Object of class \code{xprts}
}
