% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{xprt_all.xprts}
\alias{xprt_all.xprts}
\title{Export all objects from Export Object of class \code{xprts}}
\usage{
\method{xprt_all}{xprts}(obj, path = getwd(), is_xprt = FALSE)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path for exported file.}

\item{is_xprt}{TRUE = Export objects, FALSE = Don't export (default).}
}
\value{
\code{obj} invisibly.
}
\description{
Export all objects from Export Object of class \code{xprts}
}
