% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{xprt_classes}
\alias{xprt_classes}
\title{Get the Names and Classes from an Export Object}
\usage{
xprt_classes(obj)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}
}
\value{
Datafame with names and classes.
}
\description{
Get the Names and Classes from an Export Object
}
