% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{xprt_ggplot.xprts}
\alias{xprt_ggplot.xprts}
\title{Save GGPlot to PNG}
\usage{
\method{xprt_ggplot}{xprts}(
  obj,
  path = getwd(),
  width = height * 16/9,
  height = 20,
  units = "cm"
)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path used for file.}

\item{width}{Width of plot.}

\item{height}{Height of plot.}

\item{units}{UOM used for width and height.}
}
\value{
\code{obj} invisibly.
}
\description{
Save GGPlot to PNG
}
