% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{xprt_gt}
\alias{xprt_gt}
\title{Save GT to HTML}
\usage{
xprt_gt(obj, path, verbose, inline_css)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path used for file.}

\item{verbose}{TRUE = inform on every exported object.}

\item{inline_css}{See \code{gt::gtsave}. Must absolutely be set to TRUE
to include the gt table in a markdown document. Don't touch it unless you
know what you are doing.}
}
\value{
\code{obj} invisibly.
}
\description{
Save GT to HTML
}
