% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{xprt_gt.xprts}
\alias{xprt_gt.xprts}
\title{Save GT to HTML}
\usage{
\method{xprt_gt}{xprts}(obj, path = getwd(), inline_css = TRUE)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path used for file.}

\item{inline_css}{See \code{gt::gtsave}. Must absolutely be set to TRUE
to include the gt table in a markdown document. Don't touch it unless you
know what you are doing.}
}
\value{
\code{obj} invisibly.
}
\description{
Save GT to HTML
}
