% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{xprt_plotly}
\alias{xprt_plotly}
\title{Save Plotly to QS}
\usage{
xprt_plotly(obj, path)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path used for file.}
}
\value{
\code{obj} invisibly.
}
\description{
Save Plotly to QS
}
