% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprts.R
\name{xprt_plotly.xprts}
\alias{xprt_plotly.xprts}
\title{Save Plotly to QS}
\usage{
\method{xprt_plotly}{xprts}(obj, path = getwd(), verbose = FALSE)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path used for file.}

\item{verbose}{TRUE = inform on every exported object.}
}
\value{
\code{obj} invisibly.
}
\description{
Save Plotly to QS
}
