% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_acctg.R
\name{exp_acctg}
\alias{exp_acctg}
\title{Exponentiation Using Accounting Rule}
\usage{
exp_acctg(x, base = 10L)
}
\arguments{
\item{x}{Numerical vector.}

\item{base}{Positive number >= 2. Default value is 10.}
}
\value{
Numeric vector.
}
\description{
Exponentiation using accounting rule.
}
\details{
Exponentiate the number whose absolute value is greater than of equal to
the base and change the sign of the result to be the sae as the input. If
the absolute value of the number is less than the base, then multiply it
by the base. This function is the inverse of \code{log_acctg}.
}
\examples{
x <- c(-2, -1, -1/10, -0.1/10, 0, 0.1/10, 1/10, 1, 2)
y <- exp_acctg(x)
stopifnot(identical(y, c(-100, -10, -1, -0.1, 0, 0.1, 1, 10, 100)))
}
\seealso{
log_acctg
}
