% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_acctg.R
\name{log_acctg}
\alias{log_acctg}
\title{Signed \code{log1p} With a Base}
\usage{
log_acctg(x, base = 10L)
}
\arguments{
\item{x}{Numerical vector.}

\item{base}{Positive number >= 2. Default value is 10.}
}
\value{
Numeric vector.
}
\description{
Signed \code{log1p} with a base.
}
\details{
Compute \code{log_acctg(x, base) = sign(x) * log1p(abs(x)) / log(base)}.
This function is the inverse of \code{exp_acctg}.
}
\examples{
x <- c(-100, -10, -1, -0.1, 0, 0.1, 1, 10, 100)
y <- log_acctg(x)
# this test if the inverse is ok
stopifnot(isTRUE(all.equal(x, eflTools::exp_acctg(y))))
}
\seealso{
exp_acctg
}
