% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{xprt_ggplot}
\alias{xprt_ggplot}
\title{Save GGPlot to PNG}
\usage{
xprt_ggplot(obj, path, verbose, width = height, height, units)
}
\arguments{
\item{obj}{Object of class \emph{xprts}.}

\item{path}{Path used for file.}

\item{verbose}{TRUE = inform on every exported object.}

\item{width}{Width of plot.}

\item{height}{Height of plot.}

\item{units}{UOM used for width and height.}
}
\value{
\code{obj} invisibly.
}
\description{
Save GGPlot to PNG
}
